[Fixes #15691] .form-control-static changes height when empty
[Fixes #15691] .form-control-static changes height when empty by setting the minimum height to the size it is when text is in the `.form-control-static`

JSFiddle: http://jsfiddle.net/4pdo4yzo/

Compute the minimum height

@line-height-computed = ~20px
@font-size-base = 14px

This should come to approx. 34px

Add min-height to from-group-* per @cvrebert

I had this wrong. The minimum height would depend on the font-size of that specific group, thus we need to declare it once for font-size-small and once also for font-size-large