Remove half-assed inclusion of link URLs (hrefs) when printing
* Relative URLs were output as-is, which is suboptimal or even confusing for end-users
* Ditto for fragment identifiers
* Outputting long URLs inline with prose isn't good UX

JavaScript could potentially help with this (e.g. http://alistapart.com/article/improvingprint).
However, we're not currently interested in trying to tackle the hard problem of fancy print stylesheets/views
(particularly given the level of cross-browser inconsistency when it comes to printing).
So let's just keep things simple, vanilla, and unsurprising,
which should also make it easy for others to add their own print fanciness on top of Bootstrap.