Remove About > Accessibility
Currently, there is no actual accessibility page, so this avoids a 404
in the docs. More fundamentally, though, it makes more sense for
accessibility advice to be directly integrated in the docs (which we're
doing, for the most part - though we should ensure that we do
cover/mention everything where appropriate), so there's no need for a
separate accessibility page.