Link to package.json Fixed
The link to package.json in the v4 docs was linking to package.json in the v3 source.  Fixed to point to the correct v4 location.  A change like this will need to be made for all branches on version 4 if you want the docs to be linked correctly (github may be smart enough to follow the link to the current branch if you are on github but it was wrong from the main site).